# apps/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Vigi API",
        "contact": {}
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/storage/presigned-url": {
            "post": {
                "security": [
                    {
                        "JwtAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Storage"
                ],
                "summary": "Get presigned URL for file upload",
                "parameters": [
                    {
                        "description": "Request data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/storage.PresignedURLRequestDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/utils.ApiResponse-storage_PresignedURLResponseDto"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/utils.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "storage.PresignedURLRequestDto": {
            "type": "object",
            "required": [
                "contentType",
                "filename",
                "type"
            ],
            "properties": {
                "contentType": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "type": {
                    "description": "To organize files in folder structure",
                    "type": "string",
                    "enum": [
                        "user",
                        "organization"
                    ]
                }
            }
        },
        "storage.PresignedURLResponseDto": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "publicUrl": {
                    "description": "This might vary based on your S3 setup (CDN, public bucket, etc.)",
                    "type": "string"
                },
                "uploadUrl": {
                    "type": "string"
                }
            }
        },
...
        "utils.APIError": {
            "type": "object",
            "required": [
                "message",
                "status"
            ],
            "properties": {
                "message": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "utils.ApiResponse-storage_PresignedURLResponseDto": {
            "type": "object",
            "required": [
                "data",
                "message"
            ],
            "properties": {
                "data": {
                    "$ref": "#/definitions/storage.PresignedURLResponseDto"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API key authentication (pk_ prefix format)",
            "type": "apiKey",
            "name": "X-API-Key",
            "in": "header"
        },
        "JwtAuth": {
            "description": "JWT token authentication (Bearer token format)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        },
        "OrgIdAuth": {
            "description": "Organization ID authentication (header format)",
            "type": "apiKey",
            "name": "X-Organization-ID",
            "in": "header"
        }
    }
}